+++ CHANGES.md
@@ -13,6 +13,7 @@
 
 ##### Additions :tada:
 
+- The `CesiumVectorOverlays::VectorTilesRasterOverlay` now supports styling via specifying a `VectorStylingProvider` through the `pStylingProvider` parameter in `VectorTilesRasterOverlayOptions`.
 - Added `CesiumGltf::Accessor::getComponentTypeAsString` to get the corresponding semantic string for an integer component type.
 
 ### v0.64.0 - 2026-09-01